fix bugs, add ld st sti
