+++ src_vm/Makefile
@@ -17,6 +17,13 @@ SRC =   vm.c \
         init_ncurses.c \
         ncurses_windows.c \
         op.c \
+        args_functions.c \
+        commands.c \
+        add_players.c \
+        error_managment.c \
+        flags_parser.c \
+        usage.c \
+
 
 SRCDIR = ./
 
@@ -39,11 +46,9 @@ $(NAME): $(OBJ)
 	gcc $(FLAGS) -o $(NAME) $(OBJ) ../new_libft/libft.a -lncurses
 
 clean:
-	make -C ../new_libft clean
 	/bin/rm -rf $(OBJDIR)
 
 fclean: clean
-	make -C ../new_libft fclean
 	/bin/rm -f $(NAME)
 
 re: fclean all
