+++ src_vm/error_managment.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "wm.h"
+#include "vm.h"
 
 /*
 ** Error's list
