+++ src_vm/flags_parser.c
@@ -10,9 +10,9 @@
 /*                                                                            */
 /* ***************************************** ********************************* */
 
-#include "wm.h"
+#include "flags_parser.h"
 
-void	save_dump(t_data *data, char *num, int *i, int argc)
+void			save_dump(t_data *data, char *num, int *i, int argc)
 {
 	int			tmp;
 	int			j;
@@ -34,13 +34,27 @@ void	save_dump(t_data *data, char *num, int *i, int argc)
 	(*i)++;
 }
 
-void	parse_flags(t_data *data, int argc, char **argv)
+static void	print_parse(t_data *data)
 {
-	int		i;
+//	ft_printf("dump:%d\n", data->dump);
+	int	i = -1;
+	ft_printf("bot count: %d\n", data->bots_count);
+	while (++i < 4)
+	{
+		ft_printf ("%s\n", data->players[i]);
+	}
+}
+
+int				parse_flags(t_data *data, int argc, char **argv)
+{
+	int			i;
 
 	i = 1;
 	if (!ft_strcmp("-h", argv[i]) || !ft_strcmp("--help", argv[i]))
+	{
 		usage();
+		return (1);
+	}
 	if (!ft_strcmp("-dump", argv[i]))
 		save_dump(data, argv[i + 1], &i, argc);
 	else if (!ft_strcmp("-v", argv[i]))
@@ -52,6 +66,8 @@ void	parse_flags(t_data *data, int argc, char **argv)
 		add_wn(data, &i, argv, argc);
 	else
 		add_player(data, &i, argv, argc);
-	if (data->bot_cnt > MAX_PLAYERS)
+	if (data->bots_count > MAX_PLAYERS)
 		exit_error(data, 8);
+	print_parse(data);
+	return (0);
 }
